requirements.txt
appdirs==1.4.3
backports-abc==0.5
certifi==2017.4.17
matrix_io-proto==0.0.10
packaging==16.8
protobuf==3.3.0
pyparsing==2.2.0
pyzmq==16.0.2
singledispatch==3.4.0.3
six==1.10.0
tornado==4.5.1
paho-mqtt=1.4.0

